[FIX] PR 수정사항 - 도메인 디렉토리 수정 - 메소드들 delete(), restore()로 명명 - BaseTimeEntity 에서 접근제한 private으로 수정 - 누락됐던 탈퇴처리 로직 재추가 - 마이그레이션 파일 V8로 변경
