+++ src/main/java/io/dev/jobprep/domain/essentialMaterial/application/dto/EssentialMaterialService.java
@@ -1,0 +1,4 @@
+package io.dev.jobprep.domain.essentialMaterial.application.dto;
+
+public class EssentialMaterialService {
+}
