+++ src/main/java/io/dev/jobprep/domain/essentialMaterial/domain/EssentialMaterial.java
@@ -1,0 +1,4 @@
+package io.dev.jobprep.domain.essentialMaterial.domain;
+
+public class EssentialMaterial {
+}
