+++ src/main/java/io/dev/jobprep/domain/essentialMaterial/infrastructure/EssentialMaterialRepository.java
@@ -1,0 +1,4 @@
+package io.dev.jobprep.domain.essentialMaterial.infrastructure;
+
+public interface EssentialMaterialRepository {
+}
