+++ src/main/java/io/dev/jobprep/domain/essentialMaterial/presentation/dto/EssentialMaterialController.java
@@ -1,0 +1,4 @@
+package io.dev.jobprep.domain.essentialMaterial.presentation.dto;
+
+public class EssentialMaterialController {
+}
