+++ src/main/java/io/dev/jobprep/domain/users/application/dto/res/DeleteUserAccountResponse.java
@@ -1,0 +1,17 @@
+package io.dev.jobprep.domain.users.application.dto.res;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+@Builder
+public final class DeleteUserAccountResponse {
+    private String message;
+
+    public static DeleteUserAccountResponse from(String message){
+        return new DeleteUserAccountResponse(message);
+    }
+}
