+++ src/main/java/io/dev/jobprep/domain/users/domain/UserRole.java
@@ -1,0 +1,5 @@
+package io.dev.jobprep.domain.users.domain;
+
+public enum UserRole {
+   ADMIN, NORMAL
+}
