+++ src/main/java/io/dev/jobprep/domain/users/infrastructure/UserRepository.java
@@ -1,0 +1,13 @@
+package io.dev.jobprep.domain.users.infrastructure;
+
+import io.dev.jobprep.domain.users.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findUserById(Long id);
+    Optional<User> findUserByEmail(String email);
+}
