--- src/main/java/io/dev/jobprep/essentialMaterial/application/dto/EssentialMaterialService.java
@@ -1,4 +1,0 @@
-package io.dev.jobprep.essentialMaterial.application.dto;
-
-public class EssentialMaterialService {
-}
