--- src/main/java/io/dev/jobprep/essentialMaterial/domain/EssentialMaterial.java
@@ -1,4 +1,0 @@
-package io.dev.jobprep.essentialMaterial.domain;
-
-public class EssentialMaterial {
-}
