--- src/main/java/io/dev/jobprep/essentialMaterial/infrastructure/EssentialMaterialRepository.java
@@ -1,4 +1,0 @@
-package io.dev.jobprep.essentialMaterial.infrastructure;
-
-public interface EssentialMaterialRepository {
-}
