--- src/main/java/io/dev/jobprep/essentialMaterial/presentation/dto/EssentialMaterialController.java
@@ -1,4 +1,0 @@
-package io.dev.jobprep.essentialMaterial.presentation.dto;
-
-public class EssentialMaterialController {
-}
