--- src/main/java/io/dev/jobprep/users/application/dto/res/DeleteUserAccountResponse.java
@@ -1,19 +1,0 @@
-package io.dev.jobprep.users.application.dto.res;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@AllArgsConstructor
-@Builder
-public final class DeleteUserAccountResponse {
-    private String message;
-
-    public static DeleteUserAccountResponse from(String message){
-        return new DeleteUserAccountResponse(message);
-    }
-}
