--- src/main/java/io/dev/jobprep/users/domain/UserRole.java
@@ -1,5 +1,0 @@
-package io.dev.jobprep.users.domain;
-
-public enum UserRole {
-   ADMIN, NORMAL
-}
