--- src/main/java/io/dev/jobprep/users/infrastructure/UserRepository.java
@@ -1,13 +1,0 @@
-package io.dev.jobprep.users.infrastructure;
-
-import io.dev.jobprep.users.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findUserById(Long id);
-    Optional<User> findUserByEmail(String email);
-}
